Add breadth-first post mailing to avert out-of-order post receipt
Mail all (or select) topics by individual post, breadth-first in order to
cut down on messages that are not in order.
(the reply-to component of a mail message requires the value provided to
the recipient of the email message, and so cannot be used to fix the sequence
of a set of email sent sequentially by the same sender. Gmail assigns the
references (reply-to) elements when it has processed each new message, and
the duration of delay in between messages to prevent race conditions leading
to posts out of order is actually quite long (certainly longer than 1s,
probably longer than 10s)).

On the other hand, if we send the first message of every topic first, then
the second message of every topic, then the third message of every topic,
etc., then the only rate limitation we have is that required to not anger
Google for sending messages too quickly. 0.2s should suffice there.